Move hits classes to separate file
