+++ ydb/core/nbs/cloud/blockstore/libs/storage/partition_direct/dirty_map/ya.make
@@ -7,6 +7,7 @@ GENERATE_ENUM_SERIALIZATION(inflight_info.h)
 SRCS(
     ddisk_state.cpp
     dirty_map.cpp
+    hints.cpp
     inflight_info.cpp
     range_locker.cpp
 )
